Rewrite bot with modular architecture
