+++ requirements.txt
@@ -1,0 +1,5 @@
+aiohttp
+python-telegram-bot==21.6
+openai
+aiosqlite
+python-dotenv
